rename from `typeck` to `type_check` -- down with contractions!
