+++ src/main.rs
@@ -33,7 +33,7 @@ mod parser;
 mod parser2;
 mod tests;
 mod ty;
-mod typeck;
+mod type_check;
 mod unify;
 
 use std::{env, io};
