+++ src/type_check/ops.rs
@@ -1,8 +1,8 @@
 use crate::ty;
 use crate::ty::interners::TyInternTables;
 use crate::ty::BaseData;
-use crate::typeck::BaseTypeChecker;
-use crate::typeck::TypeCheckDatabase;
+use crate::type_check::BaseTypeChecker;
+use crate::type_check::TypeCheckDatabase;
 use crate::unify::InferVar;
 use crate::unify::Inferable;
 use generational_arena::Arena;
@@ -31,7 +31,7 @@ where
 
 impl<DB> BaseTypeChecker<'q, DB>
 where
-    DB: crate::typeck::TypeCheckDatabase,
+    DB: crate::type_check::TypeCheckDatabase,
 {
     /// Enqueues a closure to execute when any of the
     /// variables in `values` are unified.
