+++ src/type_check/query_definitions.rs
@@ -11,9 +11,9 @@ use crate::ty::declaration::Declaration;
 use crate::ty::interners::{HasTyInternTables, TyInternTables};
 use crate::ty::Ty;
 use crate::ty::TypeFamily;
-use crate::typeck::BaseTypeCheckResults;
-use crate::typeck::BaseTypeChecker;
-use crate::typeck::TypeCheckDatabase;
+use crate::type_check::BaseTypeCheckResults;
+use crate::type_check::BaseTypeChecker;
+use crate::type_check::TypeCheckDatabase;
 use crate::unify::InferVar;
 use crate::unify::UnificationTable;
 use generational_arena::Arena;
